update some strategy.
